refactor: extract terminal into independent feature module
- Move terminal from host-manage subdirectory to pages/terminal/
- Extract pane-registry as independent module (lib/pane-registry.ts)
- Split terminal God Component into TerminalHeader + TerminalTabBar
- Add TerminalErrorBoundary for per-pane crash isolation
- Merge tabs/activeTabId into single atomic state (fix sync issues)
- Use stateRef for stale closure safety in event handlers
- Switch useEffect → useLayoutEffect for container mount timing
- Replace Date.now()+counter with crypto.randomUUID() for pane IDs
- Trigger fit() on remaining panes after split/close via rAF
- Adjacent pane focus on close instead of always jumping to first
