+++ frontend/src/pages/terminal/components/Terminal.tsx
