+++ frontend/src/pages/terminal/components/types.ts
