+++ frontend/src/pages/terminal/constants/themes.ts
