+++ frontend/src/pages/terminal/hooks/use-host-search.ts
