+++ frontend/src/pages/terminal/hooks/use-split-pane-shortcuts.ts
